backend/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
cryptography==43.0.1
python-dotenv==1.0.1
